Feature arturo

+++ src/main/java/com/example/dailityhabits/service/impl/FrequencyServiceImpl.java
@@ -1,14 +1,13 @@
 package com.example.dailityhabits.service.impl;
 
-import com.example.dailityhabits.DTO.frequency.FrequencyDTO;
-import com.example.dailityhabits.exception.notFound.FrequencyNotFoundException;
+import com.example.dailityhabits.DTO.frequency.CreateFrequencyDTO;
+import com.example.dailityhabits.DTO.frequency.ResponseFrequencyDTO;
+import com.example.dailityhabits.exception.notFound.ReminderNotFoundException;
 import com.example.dailityhabits.mapper.FrequencyMapper;
 import com.example.dailityhabits.repository.FrequencyRepository;
 import com.example.dailityhabits.service.interfaces.FrequencyService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-
-import java.util.NoSuchElementException;
 
 @Service
 @RequiredArgsConstructor
@@ -18,37 +17,41 @@ public class FrequencyServiceImpl implements FrequencyService {
     private final FrequencyMapper frequencyMapper;
 
     @Override
-    public FrequencyDTO getFrequencybyId(Long id) {
+    public ResponseFrequencyDTO getFrequencybyId(Long id) {
         if (id == null || id <= 0) {
-            throw new FrequencyNotFoundException("Invalid ID");
+            throw new ReminderNotFoundException("Invalid ID");
         }
 
         return frequencyRepository.findById(id)
-                .map(frequencyMapper::toFrequencyDTO)
-                .orElseThrow(() -> new FrequencyNotFoundException("Frequency not found"));
+                .map(frequencyMapper::toResponseFrequencyDTO)
+                .orElseThrow(() -> new ReminderNotFoundException("Frequency not found"));
     }
 
     @Override
-    public FrequencyDTO createFrequency(FrequencyDTO frequencyDTO) {
+    public ResponseFrequencyDTO createFrequency(CreateFrequencyDTO createFrequencyDTO) {
         return frequencyMapper
-                .toFrequencyDTO(frequencyRepository.save(frequencyMapper.fromFrequencyDTO(frequencyDTO)));
+                .toResponseFrequencyDTO(frequencyRepository.save(frequencyMapper.fromCreateFrequencyDTO(createFrequencyDTO)));
     }
 
     @Override
-    public FrequencyDTO updateFrequency(FrequencyDTO frequencyDTO) {
-        if (frequencyDTO.id() == null || frequencyDTO.id() <= 0) {
-            throw new FrequencyNotFoundException("Invalid ID");
+    public ResponseFrequencyDTO updateFrequency(ResponseFrequencyDTO responseFrequencyDTO) {
+        if (frequencyRepository.existsById(responseFrequencyDTO.id())) {
+            return frequencyMapper.toResponseFrequencyDTO(
+                    frequencyRepository.save(frequencyMapper.fromResponseFrequencyDTO(responseFrequencyDTO)));
+        }else{
+            throw new ReminderNotFoundException("Frequency not found");
         }
-        return frequencyMapper.toFrequencyDTO(
-                frequencyRepository.save(frequencyMapper.fromFrequencyDTO(frequencyDTO))
-        );
     }
 
     @Override
     public void deleteFrequencyById(Long id) {
         if (id == null || id <= 0) {
-            throw new FrequencyNotFoundException("Invalid ID");
+            throw new ReminderNotFoundException("Invalid ID");
         }
-        frequencyRepository.deleteById(id);
+        if (frequencyRepository.existsById(id)) {
+            frequencyRepository.deleteById(id);
+        }else {
+            throw new ReminderNotFoundException("Frequency not found");
+        }
     }
 }
