+++ src/main/java/com/example/dailityhabits/service/impl/ReminderServiceImpl.java
@@ -1,15 +1,13 @@
 package com.example.dailityhabits.service.impl;
 
-import com.example.dailityhabits.DTO.reminder.ReminderDTO;
+import com.example.dailityhabits.DTO.reminder.CreateReminderDTO;
+import com.example.dailityhabits.DTO.reminder.ResponseReminderDTO;
 import com.example.dailityhabits.exception.notFound.ReminderNotFoundException;
 import com.example.dailityhabits.mapper.ReminderMapper;
-import com.example.dailityhabits.repository.HabitRepository;
 import com.example.dailityhabits.repository.ReminderRepository;
 import com.example.dailityhabits.service.interfaces.ReminderService;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-
-import java.util.NoSuchElementException;
 
 @Service
 @RequiredArgsConstructor
@@ -19,7 +17,7 @@ public class ReminderServiceImpl implements ReminderService {
     private final ReminderMapper reminderMapper;
 
     @Override
-    public ReminderDTO getReminderById(Long id) {
+    public ResponseReminderDTO getReminderById(Long id) {
         if (id == null || id <= 0) {
             throw new ReminderNotFoundException("Invalid ID");
         }
@@ -29,18 +27,19 @@ public class ReminderServiceImpl implements ReminderService {
     }
 
     @Override
-    public ReminderDTO createReminder(ReminderDTO reminderDTO) {
+    public ResponseReminderDTO createReminder(CreateReminderDTO createReminderDTO) {
         return reminderMapper.toReminderDTO(
-                reminderRepository.save(reminderMapper.fromReminderDTO(reminderDTO)));
+                reminderRepository.save(reminderMapper.fromCreateReminderDTO(createReminderDTO)));
     }
 
     @Override
-    public ReminderDTO updateReminder(ReminderDTO reminderDTO) {
-        if (reminderDTO.id() == null || reminderDTO.id() <= 0) {
-            throw new ReminderNotFoundException("Invalid ID");
+    public ResponseReminderDTO updateReminder(ResponseReminderDTO responseReminderDTO) {
+        if (reminderRepository.existsById(responseReminderDTO.id())){
+            return reminderMapper.toReminderDTO(
+                    reminderRepository.save(reminderMapper.fromResponseReminderDTO(responseReminderDTO)));
+        }else{
+            throw new ReminderNotFoundException("Reminder not found");
         }
-        return reminderMapper.toReminderDTO(
-                reminderRepository.save(reminderMapper.fromReminderDTO(reminderDTO)));
     }
 
     @Override
@@ -48,6 +47,8 @@ public class ReminderServiceImpl implements ReminderService {
         if (id == null || id <= 0) {
             throw new ReminderNotFoundException("Invalid ID");
         }
-        reminderRepository.deleteById(id);
+        if (reminderRepository.existsById(id)){
+            reminderRepository.deleteById(id);
+        }
     }
 }
